# enhance: [10kcp] Optimize save collection target latency (#38345) (#38370)